+++ lib/ruby_native.rb
@@ -4,6 +4,7 @@ require "ruby_native/native_version"
 require "ruby_native/native_detection"
 require "ruby_native/inertia_support"
 require "ruby_native/oauth_middleware"
+require "ruby_native/tunnel_cookie_middleware"
 require "ruby_native/engine"
 
 module RubyNative